+++ protected/controllers/AdminSampleController.php
@@ -2,11 +2,7 @@
 
 class AdminSampleController extends Controller
 {
-	/**
-   * @var string the default layout for the views. Defaults to '//layouts/datasetpage'
-   * See 'protected/views/layouts/datasetpage.php'.
-   */
-	public $layout='//layouts/datasetpage';
+
 
 	/**
 	 * @return array action filters
@@ -286,7 +282,7 @@ class AdminSampleController extends Controller
 			$model->setAttributes($attrs, true);
 		}
 
-        $this->layout = 'main';
+
 		$this->loadBaBbqPolyfills = true;
 		$this->render('admin',array(
 			'model'=>$model,