+++ protected/controllers/CurationLogController.php
@@ -3,12 +3,6 @@
 class CurationLogController extends Controller
 {
 	/**
-   * @var string the default layout for the views. Defaults to '//layouts/datasetpage'
-   * See 'protected/views/layouts/datasetpage.php'.
-   */
-
-
-	/**
 	 * @return array action filters
 	 */
 	public function filters()
@@ -79,7 +73,6 @@ class CurationLogController extends Controller
 				$this->redirect(array('view','id'=>$model->id));
 		}
 
-		$this->layout = 'datasetpage';
 		$this->render('create',array(
 			'model'=>$model,
             'dataset_id'=>$id,
@@ -113,7 +106,6 @@ class CurationLogController extends Controller
 					$this->redirect(array('view','id'=>$model->id));
 			}
 
-		  $this->layout = 'datasetpage';
 			$this->render('update',array(
 				'model'=>$model,
 			));
@@ -145,7 +137,6 @@ class CurationLogController extends Controller
 	 */
 	public function actionView($id) {
 
-        $this->layout = 'datasetpage';
         $this->render('view', array(
             'model' => $this->loadModel($id),
         ));