+++ protected/controllers/DatasetLogController.php
@@ -2,11 +2,7 @@
 
 class DatasetLogController extends Controller
 {
-	/**
-   * @var string the default layout for the views. Defaults to '//layouts/datasetpage'
-   * See 'protected/views/layouts/datasetpage.php'.
-   */
-	public $layout='//layouts/datasetpage';
+
 
 	/**
 	 * @return array action filters
@@ -128,7 +124,6 @@ class DatasetLogController extends Controller
 		if(isset($_GET['DatasetLog']))
 			$model->setAttributes($_GET['DatasetLog']);
 
-		$this->layout = 'main';
 		$this->loadBaBbqPolyfills = true;
 		$this->render('admin',array(
 			'model'=>$model,